% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_configure.R
\name{auto_dark_bg}
\alias{auto_dark_bg}
\title{Generate dark background from light foreground}
\usage{
auto_dark_bg(hex)
}
\description{
Generate dark background from light foreground
}
\keyword{internal}
