% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_configure.R
\name{auto_dark_fg}
\alias{auto_dark_fg}
\title{Generate light foreground from dark background}
\usage{
auto_dark_fg(hex)
}
\description{
Generate light foreground from dark background
}
\keyword{internal}
