% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_cache.R
\name{brand_colors}
\alias{brand_colors}
\title{Get Cached Brand Colours}
\usage{
brand_colors(mode = c("light", "dark"))
}
\arguments{
\item{mode}{\code{"light"} or \code{"dark"}.}
}
\value{
Named list of hex colour values.
}
\description{
Get Cached Brand Colours
}
