% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_colors.R
\name{brand_pal_seq}
\alias{brand_pal_seq}
\title{Sequential Palette}
\usage{
brand_pal_seq(type = "warm", n = 9, reverse = FALSE, mode = "light")
}
\arguments{
\item{type}{\code{"warm"}, \code{"cool"}, or \code{"green"}.}

\item{n}{Number of colours (default 9).}

\item{reverse}{Flip direction.}

\item{mode}{\code{"light"} or \code{"dark"}.}
}
\value{
Character vector of hex codes.
}
\description{
Sequential Palette
}
