% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_configure.R
\name{existing_brand_yml}
\alias{existing_brand_yml}
\title{Path to an existing \verb{_brand.yml} in \code{path}, or \code{NULL}}
\usage{
existing_brand_yml(path)
}
\description{
Path to an existing \verb{_brand.yml} in \code{path}, or \code{NULL}
}
\keyword{internal}
