% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_configure.R
\name{match_font_pair}
\alias{match_font_pair}
\title{Name of the font pairing matching this base/heading combo, else \code{"Custom"}}
\usage{
match_font_pair(base, heading, font_pairs)
}
\description{
Name of the font pairing matching this base/heading combo, else \code{"Custom"}
}
\keyword{internal}
