% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift_color}
\alias{shift_color}
\title{Lighten or Darken a Hex Colour}
\usage{
shift_color(hex, amount)
}
\arguments{
\item{hex}{Hex colour string.}

\item{amount}{Positive = lighten, negative = darken. Range roughly -1 to 1.}
}
\description{
Thin wrapper around colorspace for palette derivation.
}
\keyword{internal}
