% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand_pages.R
\name{brand_page_fluid}
\alias{brand_page_fluid}
\title{Branded Fluid Page}
\usage{
brand_page_fluid(
  ...,
  title = NULL,
  dark_mode = TRUE,
  dark_mode_id = "dark_mode",
  theme_args = list()
)
}
\arguments{
\item{...}{UI elements.}

\item{title}{Page title.}

\item{dark_mode}{Include a dark-mode toggle? Default \code{TRUE}.}

\item{dark_mode_id}{ID for the toggle widget.}

\item{theme_args}{Extra args passed to \code{brand_theme()}.}
}
\value{
A Shiny UI definition.
}
\description{
Drop-in replacement for \code{bslib::page_fluid()}.
}
